ImagePlugin is no longer used
