+++ core/sbt-cloudflow/src/main/scala/cloudflow/sbt/CloudflowBasePlugin.scala
@@ -23,6 +23,7 @@ import sbt.Keys._
 import sbtdocker._
 import sbtdocker.DockerKeys._
 import com.typesafe.sbt.packager.archetypes._
+import scala.util.control._
 
 import cloudflow.sbt.CloudflowKeys._
 
@@ -133,4 +134,20 @@ object CloudflowBasePlugin extends AutoPlugin {
     log.info("Successfully built and published the following image:")
     log.info(s"  $imagePushed")
   }
+}
+
+case object DockerRegistryNotSet extends Exception(DockerRegistryNotSetError.msg) with NoStackTrace with sbt.FeedbackProvidedException
+object DockerRegistryNotSetError {
+  val msg =
+    """
+              |Please set the `cloudflowDockerRegistry` sbt setting in your build.sbt file to the registry that you want to push the image to. This Docker registry must be configured for image pulling on your target Kubernetes clusters and you should `docker login` to it before building and pushing any images.
+              |Example:
+              |
+              |lazy val myProject = (project in file("."))
+              |  .enablePlugins(CloudflowAkkaStreamsApplicationPlugin)
+              |  .settings(
+              |   cloudflowDockerRegistry := Some("docker-registry-default.cluster.example.com"),
+              |   // other settings
+              |  )
+            """.stripMargin
 }
